+++ src/lorenz.h
@@ -17,7 +17,7 @@ public:
   void derivs(double time,
 	      std::vector<double>::const_iterator y,
 	      std::vector<double>::iterator dydt);
-  unsigned int size() const { return 3; }
+  size_t size() const { return 3; }
   
   std::vector<double> r_derivs(double time, 
 			       std::vector<double> y);
